Issue #37 - Encounters should track follow ups
The Contact class now has a boolean property, needsFollowUp, that is set
by new Encounters that indicate a follow up is required. An Encounter
tracks whether the contact should be followed up by keeping a boolean
property, requiresFollowUp. If the new Encounter is also the most recent
Encounter, then the Contact's needsFollowUp flag will be updated;
otherwise, the Encounter will not affect this Contact field.

+++ smwc/src/main/java/edu/usm/service/impl/EncounterServiceImpl.java
@@ -1,5 +1,6 @@
 package edu.usm.service.impl;
 
+import edu.usm.domain.Committee;
 import edu.usm.domain.Contact;
 import edu.usm.domain.Encounter;
 import edu.usm.dto.EncounterDto;
@@ -9,6 +10,8 @@ import edu.usm.service.ContactService;
 import edu.usm.service.EncounterService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Set;
 
 /**
  * Created by Andrew on 7/16/2015.
@@ -27,26 +30,56 @@ public class EncounterServiceImpl extends BasicService implements EncounterServi
         return encounterDao.findOne(id);
     }
 
-    @Override
-    public void deleteEncounter(Encounter encounter) {
-        contactService.removeEncounter(encounter.getContact(), encounter);
-        contactService.removeInitiator(encounter.getInitiator(), encounter);
-        encounterDao.delete(encounter);
+    private Set<Encounter> findAll() {
+        return encounterDao.findAll();
     }
 
     @Override
+    public void deleteEncounter(Encounter encounter) {
+        if (null != encounter.getContact()) {
+            contactService.removeEncounter(encounter.getContact(), encounter);
+        }
+
+        if (null != encounter.getInitiator()) {
+            contactService.removeInitiator(encounter.getInitiator(), encounter);
+        }
+
+        encounterDao.delete(encounter);
+    }
+
+    /*
+    * Updates the core fields of the Encounter. If this is the Contact's most recent encounter,
+    * will also update that Contact's needsFollowUp and assessment fields
+    */
+    @Override
     public void updateEncounter(Encounter existingEncounter, EncounterDto dto) {
+
+        if (null == dto) {
+            encounterDao.save(existingEncounter);
+            return;
+        }
+
         Contact initiator = contactService.findById(dto.getInitiatorId());
         existingEncounter.setAssessment(dto.getAssessment());
         existingEncounter.setNotes(dto.getNotes());
         existingEncounter.setEncounterDate(dto.getEncounterDate());
         existingEncounter.setType(dto.getType());
         existingEncounter.setInitiator(initiator);
+        existingEncounter.setRequiresFollowUp(dto.requiresFollowUp());
         encounterDao.save(existingEncounter);
+
+        Contact contact = existingEncounter.getContact();
+        if (existingEncounter == contact.getEncounters().first()) {
+            contactService.updateNeedsFollowUp(contact, existingEncounter.requiresFollowUp());
+        }
+
+        contactService.updateAssessment(contact, contactService.getUpdatedAssessment(contact));
+
     }
 
     @Override
-    public void updateEncounter(Encounter encounter) {
-        encounterDao.save(encounter);
+    public void deleteAll() {
+        Set<Encounter> encounters = findAll();
+        encounters.stream().forEach(this::deleteEncounter);
     }
 }
